Revert "dto class 자료형 변경"
This reverts commit 0d24c7f4abac7ba2893be7d02f8a8ebb78d6162d.

+++ src/main/java/com/yet/project/domain/user/SocialLogin.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class SocialLogin {
-    Long socialId;
+    Integer socialId;
     String name;
     String nameKor;
 }
